// 6_oy_11_dars_uy_ishi/dto/auth.dto.ts
export interface CreateAuthDto {
    username: string;
    email: string;
    password: string;
    role: string;
    otp: string;
    isVerified: boolean;
    otptime: string
}

export interface CreateAuthVerifyDto {
    email: string;
    otp: string;
}

export interface CreateAuthLoginDto {
    email: string;
    password: string;
}

export interface CreateAuthResendOtpyDto {
    email: string;
}

export interface CreateAuthForgotPasswordDto {
    email: string;
    otp: string;
    new_password: string;
}